added permissions checking and basic front end error handling for deleteItem

+++ sick-fits/backend/src/resolvers/Mutation.js
@@ -56,9 +56,16 @@ const Mutations = {
   async deleteItem(parent, args, ctx, info) {
     const where = { id: args.id };
     // find the item
-    const item = await ctx.db.query.item({ where }, `{id title}`);
+    const item = await ctx.db.query.item({ where }, `{id title user { id }}`);
     // check if they own the item or have permissions to delete this item
-    // TODO
+    const ownsItem = item.user.id === ctx.request.userId
+    const hasPermissions = ctx.request.user.permissions.some(permission => ['ADMIN', 'ITEMDELETE'].includes(permission))
+    console.log('ownsItem?', item.user.id, ctx.request.userId)
+    console.log('hasPermissions? ', hasPermissions)
+    if(!ownsItem && hasPermissions) {
+      throw new Error('You don\'t have permission to do that!');
+    }
+    
     // delete the item
     return ctx.db.mutation.deleteItem({ where }, info);
   },
